% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorseq.R
\name{colorseq}
\alias{colorseq}
\title{Applies html coloring to biological sequences}
\usage{
colorseq(seq, color_scheme = "clustal")
}
\arguments{
\item{seq}{a vector of amino acid sequences}

\item{color_scheme}{the desired color scheme. Available choices are:
\itemize{
\item "clustal": https://www.bioinformatics.nl/~berndb/aacolour.html
}}
}
\value{
vector of html colored sequences
}
\description{
Takes a vector of biological sequences and applies html coloring. Coloring
can be done according to biological properties such as biochemistry,  and
structure. The rendered html of color sequences can help identify patterns in
sequences based on these various properties.
}
\examples{
library(gt)
data(aa_seq)

# compare in gt
one_seq <- aa_seq$seq[1]
tibble(
  seq = .color_sequence(one_seq),
  seq_raw = one_seq
) |>
  gt() |>
  fmt_markdown()

# in a dplyr pipe
aa_seq |>
  mutate(seq_colored = .color_sequence(seq)) |>
  select(name, everything()) |>
  gt() |>
  fmt_markdown() |>
  tab_style(
    style = cell_text(font = 'Cascadia Code'),
    locations = cells_body(
    columns = contains('seq')
  )
)
}
