% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorseq.R
\name{.apply_style_color}
\alias{.apply_style_color}
\title{Apply style atomic function}
\usage{
.apply_style_color(target, color, type = "background")
}
\arguments{
\item{target}{input string}

\item{color}{color, can be hash}

\item{type}{color 'background' or 'text'}
}
\value{
character vector
}
\description{
Apply style atomic function
}
\examples{
\dpntrun {
# turn into tests

.apply_style_color('EKLATI', 'yellow', 'text')
.apply_style_color('EKLATI', 'yellow', 'background')
.apply_style_color('', 'yellow', 'background')
.apply_style_color('EKLATI', 'yellow', 'x') #error
}
}
