% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_seq.R
\name{fmt_seq}
\alias{fmt_seq}
\title{Color biological sequences}
\usage{
fmt_seq(
  gt_object,
  columns,
  color_scheme = "chemistry",
  font = "Courier New",
  ...
)
}
\arguments{
\item{gt_object}{An existing gt table}

\item{columns}{The columns to affect}

\item{color_scheme}{the desired color scheme. Available choices are:
\itemize{
\item "clustal": https://www.bioinformatics.nl/~berndb/aacolour.html
}}

\item{...}{Additional argument passed to}
}
\value{
a gt table
}
\description{
Color amino acids by biochemical properties like hydrophobicity, charge,
conservation, etc. Nucleotides not yet supported.
}
\examples{
library(gt)
library(tibble)
library(dplyr)
data(aa_seq)

aa_seq |> gt::gt() |> fmt_seq(seq, font = 'Consolas')
}
