% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_seq.R
\name{gt_theme_colorseq}
\alias{gt_theme_colorseq}
\title{Apply colorseq theme to gt table}
\usage{
gt_theme_colorseq(gt_object)
}
\arguments{
\item{gt_object}{gt object}
}
\value{
gt table
}
\description{
Apply colorseq theme to gt table
}
\examples{

aa_seq |> gt::gt() |> fmt_seq(seq, font = 'Consolas') |>  gt_theme_colorseq()
}
